cosmetic update
Pressure graph is insanely tall in order to deal with rrdtool quirks in parsing data. Re-centered graph based on a ~1014 hPa median average of corrected pressure by cutting off 4 hPa on the bottom. Then reduced height of graph by 5 px. Graph could be further shrunk due to the apparent nice behavior when pressure exceeds the manual bounds, which could prove handy after I plan to add a wind graph and get it all to fit on a 1080p monitor in vertical orientation.