Add Gas Resistance
Big update. Turned on the gas resistance function of the BME680, created a new RRD and started graphing it. That portion of the BME680 needs lots of burn in time to stabilize. This is not IAQ, this is just the raw resistance reading so far, not compensated for anything outside of the sensor or library reading the values. Currently the sensor has trouble outputting values for the first two cycles, but there is enough code to keep everything running until it's happy.