Update README.md
Ran into the upper limit of my gas.rrd file because of high readings from the BME680. Either my air got cleaner all of a sudden or the sensor drifted. Either way, several sources of documentation on the "normal" range of this sensor seem too tight, so I doubled the high limit before a NaN occurs.