Debug BME680 readings on service restart
When reefer service is restarted there is a huge spike on the pressure
graph that looks ugly and screws up all the data. Spent a few weeks
debugging. Turns out the sensor is not ready on the first call to read
its data because it is not heat_stable. This new code calls the function
to read the data twice before continuing on to the main loop where good
data is stored in the RRD. This has made the graph way more stable on
restarts. However, I think there is still a problem with the xff setting
on the RRD that is guessing data to insert after/before a bunch of
NaN's. I plan on testing this if I keep seeing spikes.