wow
I finally fixed the stupid negative spike bug (mostly in pressures).
I've been trying to fix this for YEARS. The solution was to time-align
the updates with the RRD steps. There is an additional check to make
sure the time-aligned epoch is greater than the last update time (to
squash the next bug I found).

Tobi Oetiker should have never allow the timestamp in 'rrdtool update'
to be "N" (as in "now"). It's too prone to errors!